add to support pool_1d, move pool_1d/2d code to pool.cpp/hpp
